vi:  Encourage writing held Touches and visiting held Files
Commit first revision
    of a 'time.sleep(0.001)' way to work as a pipe filter/ source
        by suspending TerminalPainter to write "/dev/stdout" for long enough

Report failures to read files
    even when raising exceptions before first entering the TerminalPainter

Revive the / ? g/ g? :g/ :g? Search Key entry
    that i killed lately with a tupo in the code

hunk-by-hunk details =>

-- front matter --

rewrite the Quirk of
     works as pipe filter, pipe source, or pipe drain, like the pipe drain:  ls |vi -

-- def main --

catch OSError, so as to

    solve 'bin/vi.py nonesuch' without traceback as:
    FileNotFoundError: [Errno 2] No such file or directory: 'nonesuch'

    solve 'bin/vi.py /dev/null/child' without traceback as:
    NotADirectoryError: [Errno 20] Not a directory: '/dev/null/child'

-- class TerminalFile --

first commit to hold a copy of the Bytes of a File awhile

-- class TerminalSkinVi --

rewrite in terms of TerminalFile

-- do_might_next_vi_file  # Vim :n\r -
-- def do_next_vi_file  # Vim :n!\r --

rewrite in terms of TerminalFile

-- class TerminalSkinVi --

++ def __init__ ++

rewrite in terms of TerminalFile

++ def run_vi_terminal ++

stop explaining the corners of no write to "/dev/stdout"

++ def do_say_more  # Vim ⌃G ++

keep on defaulting to show only the 'os.path.basename' no more of the path,
    but now call for it explicitly

add mention of the count of Bytes Touched
    only when nonzero

++ def do_might_flush_quit_vi  # Vim :wq\r ++
++ def do_flush_quit_vi  # Vim ZZ  # Vim :wq! ++
++ def do_might_flush_vi(self):  # Vim :w\r ++
++ def do_flush_vi(self):  # Vim :w!\r ++
++ def do_might_quit_vi(self):  # Vim :q\r ++
++ def do_quit_vi(self):  # Vim ZQ  # Vim :q!\r ++

rewrite in terms of TerminalFile

keep on telling ZQ and :q! to capture and print the last Python Traceback
    but tell ZZ :wq :wq! to Not print it

++ def might_keep_changes ++
++ def might_keep_files ++

first commit of these two checks to require '!' or uppercase ZZ/ZQ
    to drop Touches held, or to drop Files held

-- class TerminalEditor --

++ def __init__ ++
++ def _reinit_with_held_file_ ++

rewrite in terms of TerminalFile

++ def run_keyboard ++

stop wrongly losing exceptions caught before first entering the TerminalPainter
    instead, re-raise them immediately
        so they come out as tracebacks printed to Stderr and Sys Exit code 1